feat: modularize codebase and update examples
