+++ example/RTL/index.html
@@ -18,7 +18,7 @@
 
     <link href="../assets/css/bootstrap-responsive.min.css" rel="stylesheet">
 
-    <style type="text/css"> 
+    <style type="text/css">
       a, .introjs-tooltiptext {
         font-family: 'tahoma' !important;
       }
@@ -76,6 +76,6 @@
 
       <hr>
     </div>
-    <script type="text/javascript" src="../../intro.js"></script>
+    <script type="text/javascript" src="../../dist/intro.js"></script>
   </body>
 </html>
