+++ example/auto-position/index.html
@@ -22,7 +22,7 @@
 			tour.setOption('positionPrecedence', ['left', 'right', 'top', 'bottom']);
 			tour.start();
 		}
-	
+
 	</script>
   <style>
     .jumbotron {
@@ -77,20 +77,20 @@
           <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis mollis augue a neque cursus ac blandit orci faucibus. Phasellus nec metus purus.</p>
 
         </div>
-		
+
 		<div style='height:1000px'></div>
 		<div class="span6" data-step="4" data-intro="A step we need to scroll to">
 			<h4>Section Seven</h4>
 			<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis mollis augue a neque cursus ac blandit orci faucibus. Phasellus nec metus purus.</p>
 		</div>
-		
+
 		<div style='height:1000px'></div>
 		<div class="span6" data-step="5" data-intro="We want this one to be on top" data-position="top">
 			<h4>Section Eight</h4>
 			<p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis mollis augue a neque cursus ac blandit orci faucibus. Phasellus nec metus purus.</p>
 		</div>
-		
-		
+
+
 		<div style='height:1000px'></div>
 		<div class="span6" data-step="6" data-intro="We want this one to be on the left, but it won't fit" data-position="left">
 			<h4>Section Nine</h4>
@@ -99,6 +99,6 @@
       </div>
 
       <hr>
-    <script type="text/javascript" src="../../intro.js"></script>
+    <script type="text/javascript" src="../../dist/intro.js"></script>
   </body>
 </html>
