+++ example/bootstrap/v3/index.html
@@ -668,7 +668,7 @@
     <!-- Theme JavaScript -->
     <script src="js/agency.min.js"></script>
     <!-- Introjs -->
-    <script src="../../../intro.js"></script>
+    <script src="../../../dist/intro.js"></script>
 
 </body>
