+++ example/callbacks/onbeforechange.html
@@ -66,7 +66,7 @@
 
       <hr>
     </div>
-    <script type="text/javascript" src="../../intro.js"></script>
+    <script type="text/javascript" src="../../dist/intro.js"></script>
 
     <script type="text/javascript">
       function startIntro(){
