+++ example/custom-class/index.html
@@ -19,7 +19,7 @@
       .forLastStep {
         font-weight: bold;
       }
-      .customDefault { 
+      .customDefault {
         color: gray;
       }
       .customDefault .introjs-skipbutton {
@@ -79,6 +79,6 @@
 
       <hr>
     </div>
-    <script type="text/javascript" src="../../intro.js"></script>
+    <script type="text/javascript" src="../../dist/intro.js"></script>
   </body>
 </html>
