+++ example/dynamic-start/index.html
@@ -32,96 +32,96 @@
 
       <div id="workflow" class="section group" style="text-align:center;display:block;">
         <div id="col1" class="col span_1_of_9">
-          <div id="tool1" class="section group"> 
+          <div id="tool1" class="section group">
             <img class="tool" src="img/tools/tool1.png" onclick="step('7');">
           </div>
-          <div id="tool10" class="section group"> 
+          <div id="tool10" class="section group">
             <img class="tool" src="img/tools/tool10.png" onclick="step('13');">
           </div>
-          <div class="section group"> 
+          <div class="section group">
           </div>
-        </div>     
+        </div>
         <div id="col2" class="col span_1_of_9">
-          <div id="tool2" class="section group"> 
+          <div id="tool2" class="section group">
             <img class="tool" src="img/tools/tool2.png" onclick="step('9');">
           </div>
-          <div class="section group"> 
+          <div class="section group">
             <img class="tool" src="img/tools/line.png">
           </div>
-          <div class="section group"> 
+          <div class="section group">
           </div>
         </div>
         <div id="col3" class="col span_1_of_9">
-          <div id="tool3" class="section group"> 
+          <div id="tool3" class="section group">
             <img class="tool" src="img/tools/tool3.png" onclick="step('15');">
           </div>
-          <div class="section group"> 
+          <div class="section group">
             <img class="tool" src="img/tools/line2.png">
           </div>
-          <div class="section group"> 
+          <div class="section group">
             <img src="img/tools/line4.png">
           </div>
         </div>
         <div id="col4" class="col span_1_of_9">
-          <div id="tool4" class="section group"> 
+          <div id="tool4" class="section group">
             <img class="tool" src="img/tools/tool4.png" onclick="step('18');">
           </div>
-          <div id="tool11" class="section group"> 
+          <div id="tool11" class="section group">
             <img class="tool" src="img/tools/tool11.png" onclick="step('20');">
           </div>
-          <div id="tool14" class="section group"> 
+          <div id="tool14" class="section group">
             <img class="tool" src="img/tools/tool14.png" onclick="step('22');">
           </div>
         </div>
         <div id="col5" class="col span_1_of_9">
-          <div id="tool5" class="section group"> 
+          <div id="tool5" class="section group">
             <img class="tool" src="img/tools/tool5.png" onclick="step('24');">
           </div>
-          <div id="tool12" class="section group"> 
+          <div id="tool12" class="section group">
             <img class="tool" src="img/tools/tool12.png" onclick="step('24');">
           </div>
-          <div id="tool15" class="section group"> 
+          <div id="tool15" class="section group">
             <img class="tool" src="img/tools/tool15.png" onclick="step('24');">
           </div>
         </div>
         <div id="col6" class="col span_1_of_9">
-          <div id="tool6" class="section group"> 
+          <div id="tool6" class="section group">
             <img class="tool" src="img/tools/tool6.png" onclick="step('26');">
           </div>
-          <div id="tool13" class="section group"> 
+          <div id="tool13" class="section group">
             <img class="tool" src="img/tools/tool13.png" onclick="step('26');">
           </div>
-          <div id="tool16" class="section group"> 
+          <div id="tool16" class="section group">
             <img class="tool" src="img/tools/tool16.png" onclick="step('26');">
           </div>
         </div>
         <div id="col7" class="col span_1_of_9">
-          <div id="tool7" class="section group"> 
+          <div id="tool7" class="section group">
             <img class="tool" src="img/tools/tool7.png" onclick="step('30');">
           </div>
-          <div class="section group"> 
+          <div class="section group">
             <img src="img/tools/line3.png">
           </div>
-          <div class="section group"> 
+          <div class="section group">
             <img src="img/tools/line5.png">
           </div>
         </div>
         <div id="col8" class="col span_1_of_9">
-          <div id="tool8" class="section group"> 
+          <div id="tool8" class="section group">
             <img class="tool" src="img/tools/tool8.png" onclick="step('32');">
           </div>
-          <div class="section group"> 
+          <div class="section group">
           </div>
-          <div class="section group"> 
+          <div class="section group">
           </div>
         </div>
         <div id="col9" class="col span_1_of_9">
-          <div id="tool9" class="section group"> 
+          <div id="tool9" class="section group">
             <img class="tool" src="img/tools/tool9.png" onclick="step('34');">
           </div>
-          <div class="section group"> 
+          <div class="section group">
           </div>
-          <div class="section group"> 
+          <div class="section group">
           </div>
         </div>
       </div>
@@ -131,7 +131,7 @@
       </div>
     </div>
 
-    <script type="text/javascript" src="../../intro.js"></script>
+    <script type="text/javascript" src="../../dist/intro.js"></script>
     <script type="text/javascript" src="./resources/step.js"></script>
   </body>
 </html>
