+++ example/hint/index.html
@@ -87,29 +87,29 @@
       <hr>
 
       <div class="btn-group btn-group-justified" role="group" aria-label="...">
-          <button 
+          <button
             type="button"
             role="button"
-            data-hint="Left align" 
-            data-hintPosition="top-middle" 
+            data-hint="Left align"
+            data-hintPosition="top-middle"
             data-position="bottom"
             class="btn btn-default">Left</button>
-          <button 
+          <button
             type="button"
             role="button"
-            data-hint="Middle align" 
-            data-hintPosition="top-middle" 
+            data-hint="Middle align"
+            data-hintPosition="top-middle"
             data-position="bottom"
             class="btn btn-default">Middle</button>
-          <button 
+          <button
             type="button"
             role="button"
-            data-hint="Right align" 
-            data-hintPosition="top-middle" 
+            data-hint="Right align"
+            data-hintPosition="top-middle"
             data-position="bottom"
             class="btn btn-default">Right</button>
         </div>
     </div>
-    <script type="text/javascript" src="../../intro.js"></script>
+    <script type="text/javascript" src="../../dist/intro.js"></script>
   </body>
 </html>
