+++ example/hint/withElement.html
@@ -67,6 +67,6 @@
 
       <hr>
     </div>
-    <script type="text/javascript" src="../../intro.js"></script>
+    <script type="text/javascript" src="../../dist/intro.js"></script>
   </body>
 </html>
