+++ example/html-tooltip/index.html
@@ -68,7 +68,7 @@
       <hr>
 
     </div>
-    <script type="text/javascript" src="../../intro.js"></script>
+    <script type="text/javascript" src="../../dist/intro.js"></script>
     <script type="text/javascript">
       function startIntro(){
         var intro = introJs();
