+++ example/multi-page/index.html
@@ -61,7 +61,7 @@
       </div>
 
     </div>
-    <script type="text/javascript" src="../../intro.js"></script>
+    <script type="text/javascript" src="../../dist/intro.js"></script>
     <script type="text/javascript">
       document.getElementById('startButton').onclick = function() {
         introJs().setOption('doneLabel', 'Next page').start().oncomplete(function() {
