+++ example/multi-page/second.html
@@ -65,7 +65,7 @@
 
 
     </div>
-    <script type="text/javascript" src="../../intro.js"></script>
+    <script type="text/javascript" src="../../dist/intro.js"></script>
     <script type="text/javascript">
       if (RegExp('multipage', 'gi').test(window.location.search)) {
         introJs().start();
