+++ example/programmatic/hint.html
@@ -67,7 +67,7 @@
 
       <hr>
     </div>
-    <script type="text/javascript" src="../../intro.js"></script>
+    <script type="text/javascript" src="../../dist/intro.js"></script>
     <script type="text/javascript">
       function addHints(){
         intro = introJs();
