+++ example/programmatic/index.html
@@ -68,13 +68,13 @@
       <hr>
 
     </div>
-    <script type="text/javascript" src="../../intro.js"></script>
+    <script type="text/javascript" src="../../dist/intro.js"></script>
     <script type="text/javascript">
       function startIntro(){
         var intro = introJs();
           intro.setOptions({
             steps: [
-              { 
+              {
                 intro: "Hello world!"
               },
               {
