+++ example/svg/d3.htm
@@ -44,7 +44,7 @@
       </svg>
 
     </div>
-    <script type="text/javascript" src="../../intro.js"></script>
+    <script type="text/javascript" src="../../dist/intro.js"></script>
     <script type="text/javascript">
       document.getElementById('startButton').onclick = function() {
         introJs().start();
