+++ src/core/onResize.js
@@ -1,0 +1,5 @@
+import refresh from './refresh';
+
+export default function onResize() {
+  refresh.call(this);
+}
