+++ src/util/checkRight.js
@@ -1,0 +1,14 @@
+/**
+ * Set tooltip left so it doesn't go off the right side of the window
+ *
+ * @return boolean true, if tooltipLayerStyleLeft is ok.  false, otherwise.
+ */
+export default function checkRight(targetOffset, tooltipLayerStyleLeft, tooltipOffset, windowSize, tooltipLayer) {
+  if (targetOffset.left + tooltipLayerStyleLeft + tooltipOffset.width > windowSize.width) {
+    // off the right side of the window
+    tooltipLayer.left = `${windowSize.width - tooltipOffset.width - targetOffset.left}px`;
+    return false;
+  }
+  tooltipLayer.left = `${tooltipLayerStyleLeft}px`;
+  return true;
+}
